# docs/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "EarnPearls API",
    "description": "Version 1 API for authentication, surveys, wallet, withdrawals, and administration.",
    "version": "0.1.0"
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "ep_session",
        "description": "Opaque, revocable HttpOnly session cookie."
      },
      "csrfToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token",
        "description": "Session-bound token required for authenticated state changes."
      },
      "providerSignature": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Provider-Signature",
        "description": "Provider-adapter-specific webhook signature."
      }
    },
...
  },
  "paths": {
...
    "/withdrawals/": {
      "get": {
        "tags": [
          "Withdrawals"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "required": [
                      "id",
                      "methodCode",
                      "amount",
                      "fee",
                      "status",
                      "destinationMasked",
                      "requestedAt",
                      "updatedAt",
                      "processedAt",
                      "rejectionReason"
                    ],
                    "properties": {
                      "id": {
                        "format": "uuid",
                        "type": "string"
                      },
                      "methodCode": {
                        "type": "string"
                      },
                      "amount": {
                        "type": "object",
                        "required": [
                          "points",
                          "usdMicros",
                          "usd"
                        ],
                        "properties": {
                          "points": {
                            "pattern": "^-?[0-9]+$",
                            "type": "string"
                          },
                          "usdMicros": {
                            "pattern": "^-?[0-9]+$",
                            "type": "string"
                          },
                          "usd": {
                            "pattern": "^-?[0-9]+\\.[0-9]{6}$",
                            "type": "string"
                          }
                        }
                      },
                      "fee": {
                        "type": "object",
                        "required": [
                          "points",
                          "usdMicros",
                          "usd"
                        ],
                        "properties": {
                          "points": {
                            "pattern": "^-?[0-9]+$",
                            "type": "string"
                          },
                          "usdMicros": {
                            "pattern": "^-?[0-9]+$",
                            "type": "string"
                          },
                          "usd": {
                            "pattern": "^-?[0-9]+\\.[0-9]{6}$",
                            "type": "string"
                          }
                        }
                      },
                      "status": {
                        "anyOf": [
                          {
                            "type": "string",
                            "enum": [
                              "requested"
                            ]
                          },
                          {
                            "type": "string",
                            "enum": [
                              "under_review"
                            ]
                          },
                          {
                            "type": "string",
                            "enum": [
                              "approved"
                            ]
                          },
                          {
                            "type": "string",
                            "enum": [
                              "processing"
                            ]
                          },
                          {
                            "type": "string",
                            "enum": [
                              "paid"
                            ]
                          },
                          {
                            "type": "string",
                            "enum": [
                              "rejected"
                            ]
                          },
                          {
                            "type": "string",
                            "enum": [
                              "cancelled"
                            ]
                          }
                        ]
                      },
                      "destinationMasked": {
                        "type": "string"
                      },
                      "requestedAt": {
                        "format": "date-time",
                        "type": "string"
                      },
                      "updatedAt": {
                        "format": "date-time",
                        "type": "string"
                      },
                      "processedAt": {
                        "anyOf": [
                          {
                            "format": "date-time",
                            "type": "string"
                          },
                          {
                            "type": "null"
                          }
                        ]
                      },
                      "rejectionReason": {
                        "anyOf": [
                          {
                            "type": "string"
                          },
                          {
                            "type": "null"
                          }
                        ]
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Withdrawals"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "methodCode",
                  "points",
                  "destination"
                ],
                "properties": {
                  "methodCode": {
                    "minLength": 1,
                    "maxLength": 64,
                    "type": "string"
                  },
                  "points": {
                    "pattern": "^[1-9][0-9]*$",
                    "type": "string"
                  },
                  "destination": {
                    "minLength": 3,
                    "maxLength": 500,
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "minLength": 16,
              "maxLength": 200,
              "type": "string"
            },
            "in": "header",
            "name": "idempotency-key",
            "required": true
          }
        ],
        "security": [
          {
            "cookieAuth": [],
            "csrfToken": []
          }
        ],
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "id",
                    "methodCode",
                    "amount",
                    "fee",
                    "status",
                    "destinationMasked",
                    "requestedAt",
                    "updatedAt",
                    "processedAt",
                    "rejectionReason"
                  ],
                  "properties": {
                    "id": {
                      "format": "uuid",
                      "type": "string"
                    },
                    "methodCode": {
                      "type": "string"
                    },
                    "amount": {
                      "type": "object",
                      "required": [
                        "points",
                        "usdMicros",
                        "usd"
                      ],
                      "properties": {
                        "points": {
                          "pattern": "^-?[0-9]+$",
                          "type": "string"
                        },
                        "usdMicros": {
                          "pattern": "^-?[0-9]+$",
                          "type": "string"
                        },
                        "usd": {
                          "pattern": "^-?[0-9]+\\.[0-9]{6}$",
                          "type": "string"
                        }
                      }
                    },
                    "fee": {
                      "type": "object",
                      "required": [
                        "points",
                        "usdMicros",
                        "usd"
                      ],
                      "properties": {
                        "points": {
                          "pattern": "^-?[0-9]+$",
                          "type": "string"
                        },
                        "usdMicros": {
                          "pattern": "^-?[0-9]+$",
                          "type": "string"
                        },
                        "usd": {
                          "pattern": "^-?[0-9]+\\.[0-9]{6}$",
                          "type": "string"
                        }
                      }
                    },
                    "status": {
                      "anyOf": [
                        {
                          "type": "string",
                          "enum": [
                            "requested"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "under_review"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "approved"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "processing"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "paid"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "rejected"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "cancelled"
                          ]
                        }
                      ]
                    },
                    "destinationMasked": {
                      "type": "string"
                    },
                    "requestedAt": {
                      "format": "date-time",
                      "type": "string"
                    },
                    "updatedAt": {
                      "format": "date-time",
                      "type": "string"
                    },
                    "processedAt": {
                      "anyOf": [
                        {
                          "format": "date-time",
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "rejectionReason": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/admin/dashboard": {
      "get": {
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "users",
                    "surveys",
                    "withdrawals",
                    "providers"
                  ],
                  "properties": {
                    "users": {
                      "type": "object",
                      "required": [
                        "total",
                        "verified",
                        "limited"
                      ],
                      "properties": {
                        "total": {
                          "minimum": 0,
                          "type": "integer"
                        },
                        "verified": {
                          "minimum": 0,
                          "type": "integer"
                        },
                        "limited": {
                          "minimum": 0,
                          "type": "integer"
                        }
                      }
                    },
                    "surveys": {
                      "type": "object",
                      "required": [
                        "available",
                        "pendingParticipations"
                      ],
                      "properties": {
                        "available": {
                          "minimum": 0,
                          "type": "integer"
                        },
                        "pendingParticipations": {
                          "minimum": 0,
                          "type": "integer"
                        }
                      }
                    },
                    "withdrawals": {
                      "type": "object",
                      "required": [
                        "requested",
                        "reserved"
                      ],
                      "properties": {
                        "requested": {
                          "minimum": 0,
                          "type": "integer"
                        },
                        "reserved": {
                          "type": "object",
                          "required": [
                            "points",
                            "usdMicros",
                            "usd"
                          ],
                          "properties": {
                            "points": {
                              "pattern": "^-?[0-9]+$",
                              "type": "string"
                            },
                            "usdMicros": {
                              "pattern": "^-?[0-9]+$",
                              "type": "string"
                            },
                            "usd": {
                              "pattern": "^-?[0-9]+\\.[0-9]{6}$",
                              "type": "string"
                            }
                          }
                        }
                      }
                    },
                    "providers": {
                      "type": "object",
                      "required": [
                        "enabled",
                        "degraded"
                      ],
                      "properties": {
                        "enabled": {
                          "minimum": 0,
                          "type": "integer"
                        },
                        "degraded": {
                          "minimum": 0,
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/admin/users": {
      "get": {
        "tags": [
          "Admin"
        ],
        "parameters": [
          {
            "schema": {
              "maxLength": 200,
              "type": "string"
            },
            "in": "query",
            "name": "search",
            "required": false
          },
          {
            "schema": {
              "maxLength": 40,
              "type": "string"
            },
            "in": "query",
            "name": "status",
            "required": false
          },
          {
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 50,
              "type": "integer"
            },
            "in": "query",
            "name": "limit",
            "required": false
          }
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "required": [
                      "id",
                      "email",
                      "displayName",
                      "countryCode",
                      "accountStatus",
                      "emailVerified",
                      "limitTemplateId",
                      "createdAt"
                    ],
                    "properties": {
                      "id": {
                        "format": "uuid",
                        "type": "string"
                      },
                      "email": {
                        "format": "email",
                        "type": "string"
                      },
                      "displayName": {
                        "type": "string"
                      },
                      "countryCode": {
                        "type": "string"
                      },
                      "accountStatus": {
                        "type": "string"
                      },
                      "emailVerified": {
                        "type": "boolean"
                      },
                      "limitTemplateId": {
                        "anyOf": [
                          {
                            "format": "uuid",
                            "type": "string"
                          },
                          {
                            "type": "null"
                          }
                        ]
                      },
                      "createdAt": {
                        "format": "date-time",
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "/v1",
      "description": "Current origin"
    }
  ],
...
}